Computes an optimized repartition of the HPC computing

This function uses the limits of maximum concurrent jobs, the input/output time and computng time. The time are derived using profvis package.

HPC_optim(
  len,
  hpc_lim = 10000,
  io_time = 20000,
  computing_time = 1830,
  AssocGrpCpuLimit = 258,
  pl = TRUE,
  core_per_task = 1
)

Arguments

len
the number of time the computation needs to be iterated

hpc_lim
the maximum number of subjobs per user

io_time
the input/output time in ms, default to 20e3

computing_time
the computing time in ms, default to 1830

AssocGrpCpuLimit
the maximum number of CPU per user, default to 258

pl
logical, if TRUE a plot is produced, default to TRUE

core_per_task
number of CPU per task, default to 1.

Value

a data.frame with columns task_per_array, exec_time_min and n_array