Internal function to get the learners

The following learners "classif.h2o.glm", "classif.lda", "classif.mda", "classif.naiveBayes", "classif.IBk", "classif.kknn", "classif.featureless" do not get transformed into a TuneWrapper. The following learners c("classif.h2o.gbm", "classif.h2o.deeplearning", "classif.nnTrain", "classif.xgboost") are tuned with mlr::makeTuneControlRandom(). All other learners are tuned with mlr::makeTuneControlGrid(). Setting the possible hyper-parameters values is handled by get_ps()

get_learners_internal(
  lrn_ids,
  data,
  inner_resampling,
  grid_resolution,
  .info = FALSE,
  pca = FALSE,
  corr = FALSE,
  prob = TRUE,
  randomit = 100,
  mes
)

Arguments

lrn_ids
character, list of mlr learner idenfitiers

data
a named list with two elements data and labels

inner_resampling
resampleDesc from mlr, the inner folds of the nested resampling

grid_resolution
numeric, defines the granularity of the discrete tuning grid

.info
logical, controls the amount of information printed when tuning

pca
logical, is a PCA performed

corr
logical, are highly correlated predictors removed

prob
logical, controls the type of output, if TRUE probabilities, if FALSE response

randomit
numeric, the number of iteration for the random discrete tuning

mes
mlr list of measure to compute while tuning, the learner are tuning against the first element

Value

a list of mlr learners